pyproject.toml
[tool.poetry]
name = "Cog-Vyper-Contracts"
version = "0.1.0"
description = ""
authors = ["ControlCplusControlV"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10.6"
colorama = "^0.4.6"
titanoboa = "^0.1.7"
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
